+++ tests/inference/qmhl_loss_test.py
@@ -22,6 +22,7 @@ import sympy
 import tensorflow as tf
 import tensorflow_probability as tfp
 
+from qhbmlib.data import quantum_data
 from qhbmlib.inference import ebm
 from qhbmlib.inference import qhbm
 from qhbmlib.inference import qhbm_utils
@@ -29,7 +30,6 @@ from qhbmlib.inference import qmhl_loss
 from qhbmlib.inference import qnn
 from qhbmlib.models import circuit
 from qhbmlib.models import energy
-from qhbmlib import quantum_data
 from tests import test_util
